+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Remove warning in `Pagination`
+- Feature: Add `color` prop to `HorizontalStroke`.
+- Fix: Remove `classNames` bug from `HorizontalStroke`.
+- Fix: Remove warning in `Pagination`.
 
 ## [2.69.0] - 2020-05-12